+++ android/src/main/java/com/fluttercandies/flutter_ali_auth/mask/DecoyMaskActivity.java
@@ -19,48 +19,50 @@ import java.util.concurrent.TimeUnit;
 public class DecoyMaskActivity extends Activity {
 
 
-    public static String TAG = DecoyMaskActivity.class.getSimpleName();
+    public static final String TAG = DecoyMaskActivity.class.getSimpleName();
 
 //    public static boolean isRunning = false;
 
     @Override
     protected void onCreate(@Nullable Bundle savedInstanceState) {
 //        isRunning = true;
-
-        Log.i(TAG,"onCreate");
+        Log.i(TAG, "onCreate");
         AuthClient authClient = AuthClient.getInstance();
         AuthClient.decoyMaskActivity = this;
         // override the auth path open enter animation
-        if (authClient.getAuthModel().getAuthUIStyle() == Constant.DIALOG_PORT){
-            overridePendingTransition(R.anim.zoom_in,R.anim.stay_animation);
-        }else{
-            overridePendingTransition(R.anim.slide_up,R.anim.stay_animation);
+        if (authClient.getAuthModel().getAuthUIStyle() == Constant.DIALOG_PORT) {
+            overridePendingTransition(R.anim.zoom_in, R.anim.stay_animation);
+        } else {
+            overridePendingTransition(R.anim.slide_up, R.anim.stay_animation);
         }
         PhoneNumberAuthHelper authHelper = authClient.mAuthHelper;
-        authHelper.getLoginToken(this.getBaseContext(),authClient.getLoginTimeout());
+        authHelper.getLoginToken(this.getBaseContext(), authClient.getLoginTimeout());
         super.onCreate(savedInstanceState);
     }
 
     boolean isPause = false;
+
     @Override
     protected void onPause() {
-        Log.i(TAG,"onPause");
+        Log.i(TAG, "onPause");
         isPause = true;
         super.onPause();
     }
 
     @Override
     protected void onResume() {
-        if (isPause){
+        Log.i(TAG, "onResume");
+
+        if (isPause) {
             //TopActivityBack
-            Runnable runnable = new Runnable(){
+            Runnable runnable = new Runnable() {
                 @Override
                 public void run() {
-                   finish();
+                    finish();
                 }
             };
             ScheduledExecutorService scheduledExecutorService = Executors.newScheduledThreadPool(1);
-            scheduledExecutorService.schedule(runnable,0, TimeUnit.MILLISECONDS);
+            scheduledExecutorService.schedule(runnable, 0, TimeUnit.MILLISECONDS);
         }
         super.onResume();
     }
@@ -69,7 +71,7 @@ public class DecoyMaskActivity extends Activity {
     @Override
     public void finish() {
         super.finish();
-        Log.i(TAG,"finish");
+        Log.i(TAG, "finish");
         AuthClient.decoyMaskActivity = null;
     }